pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "langchain-community>=0.3.24",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.17",
    "openai>=1.78.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "tabulate>=0.9.0",
    "sqlalchemy>=2.0.41",
    "pandas>=2.2.3",
    "plotly>=6.1.0",
    "graphviz>=0.20.3",
]
